@@ -6,7 +6,14 @@
 
 require('./bootstrap');
 
-window.Vue = require('vue').default;
+/**
+ * Next, we will create a fresh Vue application instance.
+ * Then, you may begin adding components to this application
+ * or customize the JavaScript scaffolding to fit your unique needs.
+ */
+
+import { createApp } from 'vue'
+const app = createApp({});
 
 /**
  * The following block of code may be used to automatically register your
@@ -19,14 +26,11 @@ window.Vue = require('vue').default;
 // const files = require.context('./', true, /\.vue$/i)
 // files.keys().map(key => Vue.component(key.split('/').pop().split('.')[0], files(key).default))
 
-Vue.component('example-component', require('./components/ExampleComponent.vue').default);
+import ExampleComponent from './components/ExampleComponent.vue';
+app.component("example-component", ExampleComponent);
 
 /**
- * Next, we will create a fresh Vue application instance and attach it to
- * the page. Then, you may begin adding components to this application
- * or customize the JavaScript scaffolding to fit your unique needs.
+ * Next, we will attach the Vue instance to the page.
  */
 
-const app = new Vue({
-    el: '#app',
-});
+app.mount('#app');
